fix: Minor adjustment based on review

+++ contracts/ExecutionManager.sol
@@ -25,7 +25,7 @@ contract ExecutionManager is InheritedStrategies, NonceManager, StrategyManager,
     address public protocolFeeRecipient;
 
     // Maximum creator fee (in basis point)
-    uint256 public maximumCreatorFeeBp = 1_000;
+    uint16 public maximumCreatorFeeBp = 1_000;
 
     // Creator fee manager
     ICreatorFeeManager public creatorFeeManager;
@@ -91,7 +91,7 @@ contract ExecutionManager is InheritedStrategies, NonceManager, StrategyManager,
             // 0 --> Creator fee and adjustment of protocol fee
             if (address(creatorFeeManager) != address(0)) {
                 (recipients[1], fees[1]) = creatorFeeManager.viewCreatorFee(makerBid.collection, price, itemIds);
-                if (fees[1] * 10_000 > (price * maximumCreatorFeeBp)) revert CreatorFeeBpTooHigh();
+                if (fees[1] * 10_000 > (price * uint256(maximumCreatorFeeBp))) revert CreatorFeeBpTooHigh();
             }
 
             uint256 minTotalFee = (price * strategyInfo[makerBid.strategyId].minTotalFee) / 10_000;
@@ -113,7 +113,6 @@ contract ExecutionManager is InheritedStrategies, NonceManager, StrategyManager,
 
             // 2 --> Amount for seller
             fees[2] = price - fees[1] - fees[0];
-
             recipients[2] = takerAsk.recipient == address(0) ? sender : takerAsk.recipient;
         }
     }
@@ -147,7 +146,7 @@ contract ExecutionManager is InheritedStrategies, NonceManager, StrategyManager,
             // 0 --> Creator fee and adjustment of protocol fee
             if (address(creatorFeeManager) != address(0)) {
                 (recipients[1], fees[1]) = creatorFeeManager.viewCreatorFee(makerAsk.collection, price, itemIds);
-                if (fees[1] * 10_000 > (price * maximumCreatorFeeBp)) revert CreatorFeeBpTooHigh();
+                if (fees[1] * 10_000 > (price * uint256(maximumCreatorFeeBp))) revert CreatorFeeBpTooHigh();
             }
             uint256 minTotalFee = (price * strategyInfo[makerAsk.strategyId].minTotalFee) / 10_000;
